+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 ## [Unreleased]
 
 ### Added
+- Added `Request.messages`, `Request.errors` and `Request.results`
 - Types for `SELECT.stream()`
 - `service.actions`
 ### Changed